# AnotherMe/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnotherMe Gateway API",
    "version": "0.3.0",
    "description": "Unified backend gateway for AnotherMe Web and Mobile clients.\n\nCapabilities follow the registry defined in `agents/foundation/capability_registry.py`.\nValid capability ids: `course_generate`, `problem_video_generate`, `quiz_practice`,\n`interactive_demo`, `ai_tutor_chat`, `deep_solve`, `deep_research`,\n`math_animator`, `visualize`, `co_writer`.\n\nP1 changes:\n  - Capability endpoints (P0-3) are now wired.\n  - Global error envelope (`{ error_code, message, request_id, details? }`).\n  - Live-book / co-writer / courses routes are first-class.\n\nAuth: Phase 0 uses static `Bearer <GATEWAY_API_TOKEN>`. Phase 2 will switch to Supabase JWT.",
    "contact": {
      "name": "AnotherMe",
      "url": "https://github.com/anotherme/anotherme"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Phase 0: static token from GATEWAY_API_TOKEN env var. Phase 2: Supabase JWT."
      }
    },
    "schemas": {
...
      "ChatMessage": {
        "type": "object",
        "required": ["role", "content"],
        "properties": {
          "role": { "type": "string", "enum": ["system", "user", "assistant"] },
          "content": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": ["type", "text"],
                  "properties": {
                    "type": { "type": "string" },
                    "text": { "type": "string" }
                  }
                }
              }
            ]
          }
        }
      },
      "ChatRequest": {
        "type": "object",
        "required": ["messages", "model", "api_key"],
        "properties": {
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/ChatMessage" } },
          "model": { "type": "string", "default": "gpt-4o" },
          "api_key": { "type": "string", "default": "" },
          "base_url": { "type": "string" },
          "temperature": { "type": "number", "default": 0.1 },
          "max_tokens": { "type": "integer", "default": 4096 },
          "system_prompt": { "type": "string" },
          "capability": {
            "type": "string",
            "default": "ai_tutor_chat",
            "description": "One of: ai_tutor_chat, deep_solve, deep_research, course_generate, problem_video_generate, math_animator, visualize, co_writer, quiz_practice, interactive_demo"
          },
          "streaming": { "type": "boolean", "default": true },
          "user_id": { "type": "string", "default": "anonymous" },
          "request_id": { "type": "string" },
          "learning_context": { "type": "object", "additionalProperties": true }
        }
      },
      "ChatNonStreamingResponse": {
        "type": "object",
        "required": ["success", "assistant_text", "request_id"],
        "properties": {
          "success": { "type": "boolean" },
          "assistant_text": { "type": "string" },
          "request_id": { "type": "string" }
        }
      },
      "StreamEvent_AgentStart": {
        "type": "object",
        "required": ["type", "data"],
        "properties": {
          "type": { "type": "string", "const": "agent_start" },
          "data": {
            "type": "object",
            "required": ["messageId", "agentId", "agentName"],
            "properties": {
              "messageId": { "type": "string" },
              "agentId": { "type": "string" },
              "agentName": { "type": "string" }
            }
          }
        }
      },
      "StreamEvent_Thinking": {
        "type": "object",
        "required": ["type", "data"],
        "properties": {
          "type": { "type": "string", "const": "thinking" },
          "data": {
            "type": "object",
            "required": ["stage", "agentId", "reasoning"],
            "properties": {
              "stage": { "type": "string" },
              "agentId": { "type": "string" },
              "reasoning": { "type": "string" }
            }
          }
        }
      },
      "StreamEvent_TextDelta": {
        "type": "object",
        "required": ["type", "data"],
        "properties": {
          "type": { "type": "string", "const": "text_delta" },
          "data": {
            "type": "object",
            "required": ["content", "messageId"],
            "properties": {
              "content": { "type": "string" },
              "messageId": { "type": "string" }
            }
          }
        }
      },
      "StreamEvent_Done": {
        "type": "object",
        "required": ["type", "data"],
        "properties": {
          "type": { "type": "string", "const": "done" },
          "data": {
            "type": "object",
            "required": ["totalActions", "totalAgents", "agentHadContent"],
            "properties": {
              "totalActions": { "type": "integer" },
              "totalAgents": { "type": "integer" },
              "agentHadContent": { "type": "boolean" }
            }
          }
        }
      },
      "StreamEvent_Error": {
        "type": "object",
        "required": ["type", "data"],
        "properties": {
          "type": { "type": "string", "const": "error" },
          "data": {
            "type": "object",
            "required": ["message"],
            "properties": {
              "message": { "type": "string" }
            }
          }
        }
      },
...
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/ai/chat": {
      "post": {
        "tags": ["ai-chat"],
        "summary": "Stream chat via Server-Sent Events",
        "description": "Streams `agent_start`, `thinking`, `text_delta`, `done`/`error` events.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream of chat events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/StreamEvent_AgentStart" },
                    { "$ref": "#/components/schemas/StreamEvent_Thinking" },
                    { "$ref": "#/components/schemas/StreamEvent_TextDelta" },
                    { "$ref": "#/components/schemas/StreamEvent_Done" },
                    { "$ref": "#/components/schemas/StreamEvent_Error" }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/ai/chat/non-streaming": {
      "post": {
        "tags": ["ai-chat"],
        "summary": "Non-streaming chat (collects all events and returns final text)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Final assistant text",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChatNonStreamingResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}